Changed values to bigInteger to handle big values

+++ src/main/kotlin/com/pietri/tom/aoc/aoc2019/day9/day9.kt
@@ -1,6 +1,8 @@
 package com.pietri.tom.aoc.aoc2019.day9
 
-fun computeFirstSolution(input: String, value: Int): Int {
-    val computer = Computer(input.split(",").map { it.toInt() }.toMutableList())
+import java.math.BigInteger
+
+fun computeFirstSolution(input: String, value: Int): BigInteger {
+    val computer = Computer(input.split(",").map { it.toBigInteger() }.toMutableList())
     return computer.getNextOutput(value)
 }
